backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.1.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pyrebase4==4.7.1
firebase-admin==6.2.0
chromadb==0.4.18
langchain==0.0.335
openai==1.3.5
requests
email-validator
jose==3.3.0